+++ packages/mask/src/database/index.ts
@@ -1,5 +1,5 @@
 export * from '../../background/database/avatar-cache/avatar'
+export * from '../../background/database/post'
 export * from './Persona/helpers'
 export * from './Persona/types'
-export * from './post'
 export * from './Plugin'
