+++ packages/mask/src/extension/background-script/CryptoServices/appendShareTarget.ts
@@ -3,7 +3,7 @@ import { GunAPI as Gun2 } from '../../../network/gun/'
 import { ProfileIdentifier, PostIVIdentifier } from '../../../database/type'
 import { prepareRecipientDetail } from './prepareRecipientDetail'
 import { cryptoProviderTable } from './cryptoProviderTable'
-import { updatePostDB, RecipientDetail, RecipientReason } from '../../../database/post'
+import { updatePostDB, RecipientDetail, RecipientReason } from '../../../../background/database/post'
 import { getNetworkWorkerUninitialized } from '../../../social-network/worker'
 import { queryPrivateKey, queryLocalKey } from '../../../database'
 import { IdentifierMap } from '../../../database/IdentifierMap'
