+++ packages/mask/src/extension/background-script/CryptoServices/encryptTo.ts
@@ -6,7 +6,7 @@ import { queryPrivateKey, queryLocalKey, queryProfile } from '../../../database'
 import { ProfileIdentifier, PostIVIdentifier } from '../../../database/type'
 import { prepareRecipientDetail } from './prepareRecipientDetail'
 import { getNetworkWorker } from '../../../social-network/worker'
-import { createPostDB, PostRecord } from '../../../database/post'
+import { createPostDB, PostRecord } from '../../../../background/database/post'
 import { queryPersonaByProfileDB } from '../../../database/Persona/Persona.db'
 import { compressSecp256k1Key } from '../../../utils/type-transform/SECP256k1-Compression'
 import { i18n } from '../../../../shared-ui/locales_legacy'
