+++ packages/web3-shared/evm/package.json
@@ -7,10 +7,12 @@
     "@ensdomains/address-encoder": "^0.2.6",
     "@masknet/web3-constants": "workspace:*",
     "@masknet/web3-contracts": "workspace:*",
+    "@types/ethereum-blockies": "^0.1.0",
     "@types/punycode": "^2.1.0",
     "@unstoppabledomains/resolution": "^5.0.1",
     "bignumber.js": "9.0.1",
     "ethereum-blockies": "^0.1.1",
+    "fortmatic": "^2.2.1",
     "fuse.js": "^6.4.6",
     "lodash-es": "^4.17.21",
     "punycode": "^2.1.1",
@@ -21,8 +23,7 @@
     "wallet.ts": "^1.0.1",
     "web3": "1.5.2",
     "web3-core-helpers": "1.5.2",
-    "web3-utils": "1.5.2",
-    "fortmatic": "^2.2.1"
+    "web3-utils": "1.5.2"
   },
   "devDependencies": {
     "@types/use-subscription": "^1.0.0"
